agents/rmd/pyproject.toml
[project]
name = "rmd-agent"
version = "0.1.0"
description = "RMD Eligibility Advisor Agent"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "strands-agents[anthropic]>=1.35.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
